requirements.txt
# Google AI & Services
google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# PDF Processing
PyPDF2

# Communication Tools
twilio

# Environment & Utilities
python-dotenv